refactor: separated bookid and userid validations
