+++ middlewares/validations/book.js
@@ -1,5 +1,4 @@
 const { booksDB } = require('../../db');
-const { userDB } = require('../../db');
 
 const validateBookId = (req, res, next) => {
   const { id } = req.params;
@@ -12,18 +11,6 @@ const validateBookId = (req, res, next) => {
   next();
 };
 
-const validateUserId = (req, res, next) => {
-  const { userId } = req.params;
-
-  if (userId < 0 || userId > userDB.length) {
-    res.send('user id does not exist').status(409);
-    return;
-  }
-
-  next();
-};
-
 module.exports = {
   validateBookId,
-  validateUserId,
 };
