+++ middlewares/validations/user.js
@@ -1,0 +1,16 @@
+const { userDB } = require('../../db');
+
+const validateUserId = (req, res, next) => {
+  const { userId } = req.params;
+
+  if (userId < 0 || userId > userDB.length) {
+    res.send('user id does not exist').status(409);
+    return;
+  }
+
+  next();
+};
+
+module.exports = {
+  validateUserId,
+};
