+++ routers/user.js
@@ -5,10 +5,10 @@ const {
   createUser,
   putBooksInCart,
 } = require('../controllers/user');
-const { validateUserId } = require('../middlewares');
+const { userValidation } = require('../middlewares');
 
 userRouter.get('/find', getUserByName);
 userRouter.post('/create', createUser);
-userRouter.post('/:userId/cart', validateUserId, putBooksInCart);
+userRouter.post('/:userId/cart', userValidation.validateUserId, putBooksInCart);
 
 module.exports = userRouter;
